backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic-settings==2.5.2
httpx==0.27.2
sqlalchemy==2.0.36
psycopg2-binary==2.9.9
redis==5.1.1
qdrant-client==1.9.1
watchdog==4.0.1
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
